Ensure services have started before testing them
In certain cases Nginx may not have started after being installed.
As a result, tasks in test.yaml would fail, as would the role as a
consequence. Force the handlers to run before the end of the play is
reached and before test.yaml is imported ensures that services are
running.